requirements.txt
python-chess>=1.10.0
cairosvg>=2.7.0
Pillow>=10.0.0
zstandard>=0.22.0
